refactor common code into ezRmlUiCanvasComponentBase
