+++ Code/EnginePlugins/RmlUiPlugin/Components/Implementation/RmlUiCanvas3DComponent.cpp
@@ -20,16 +20,13 @@
 #include <RendererFoundation/Resources/Texture.h>
 
 // clang-format off
-EZ_BEGIN_COMPONENT_TYPE(ezRmlUiCanvas3DComponent, 4, ezComponentMode::Static)
+EZ_BEGIN_COMPONENT_TYPE(ezRmlUiCanvas3DComponent, 1, ezComponentMode::Static)
 {
   EZ_BEGIN_PROPERTIES
   {
     EZ_RESOURCE_ACCESSOR_PROPERTY("Mesh", GetMesh, SetMesh)->AddAttributes(new ezAssetBrowserAttribute("CompatibleAsset_Mesh_Static")),
-    EZ_RESOURCE_ACCESSOR_PROPERTY("RmlFile", GetRmlResource, SetRmlResource)->AddAttributes(new ezAssetBrowserAttribute("CompatibleAsset_Rml_UI")),
     EZ_ACCESSOR_PROPERTY("TextureSize", GetTextureSize, SetTextureSize)->AddAttributes(new ezSuffixAttribute("px"), new ezDefaultValueAttribute(ezVec2U32(512, 512)), new ezClampValueAttribute(ezVec2U32(0), ezVec2U32(4096))),
     EZ_ACCESSOR_PROPERTY("DpiScale", GetDpiScale, SetDpiScale)->AddAttributes(new ezDefaultValueAttribute(1.0f)),
-    EZ_ACCESSOR_PROPERTY("AutobindBlackboards", GetAutobindBlackboards, SetAutobindBlackboards)->AddAttributes(new ezDefaultValueAttribute(true)),
-    EZ_ACCESSOR_PROPERTY("OnDemandUpdate", GetOnDemandUpdate, SetOnDemandUpdate)->AddAttributes(new ezDefaultValueAttribute(true)),
     EZ_ACCESSOR_PROPERTY("ClearStaleInput", GetClearStaleInput, SetClearStaleInput)->AddAttributes(new ezDefaultValueAttribute(true)),
     EZ_ACCESSOR_PROPERTY("IsInteractive", IsInteractive, SetInteractive)->AddAttributes(new ezDefaultValueAttribute(true)),
   }
@@ -37,8 +34,6 @@ EZ_BEGIN_COMPONENT_TYPE(ezRmlUiCanvas3DComponent, 4, ezComponentMode::Static)
   EZ_BEGIN_MESSAGEHANDLERS
   {
     EZ_MESSAGE_HANDLER(ezMsgExtractGeometry, OnMsgExtractGeometry),
-    EZ_MESSAGE_HANDLER(ezMsgExtractRenderData, OnMsgExtractRenderData),
-    EZ_MESSAGE_HANDLER(ezMsgRmlUiReload, OnMsgReload)
   }
   EZ_END_MESSAGEHANDLERS;
   EZ_BEGIN_ATTRIBUTES
@@ -50,47 +45,36 @@ EZ_BEGIN_COMPONENT_TYPE(ezRmlUiCanvas3DComponent, 4, ezComponentMode::Static)
 EZ_END_COMPONENT_TYPE
 // clang-format on
 
-ezRmlUiCanvas3DComponent::ezRmlUiCanvas3DComponent() = default;
+ezRmlUiCanvas3DComponent::ezRmlUiCanvas3DComponent()
+{
+  m_vSize = ezVec2U32(512, 512);
+}
+
 ezRmlUiCanvas3DComponent::~ezRmlUiCanvas3DComponent() = default;
 ezRmlUiCanvas3DComponent& ezRmlUiCanvas3DComponent::operator=(ezRmlUiCanvas3DComponent&& rhs) = default;
 
-void ezRmlUiCanvas3DComponent::Initialize()
+void ezRmlUiCanvas3DComponent::SerializeComponent(ezWorldWriter& inout_stream) const
 {
-  SUPER::Initialize();
+  SUPER::SerializeComponent(inout_stream);
 
-  UpdateAutobinding();
+  ezStreamWriter& s = inout_stream.GetStream();
+
+  s << m_vSize;
+  s << m_bClearStaleInput;
+  s << m_bIsInteractive;
+  s << m_fDpiScale;
 }
 
-void ezRmlUiCanvas3DComponent::Deinitialize()
+void ezRmlUiCanvas3DComponent::DeserializeComponent(ezWorldReader& inout_stream)
 {
-  SUPER::Deinitialize();
+  SUPER::DeserializeComponent(inout_stream);
+  const ezUInt32 uiVersion = inout_stream.GetComponentTypeVersion(GetStaticRTTI());
+  ezStreamReader& s = inout_stream.GetStream();
 
-  ezGALDevice::GetDefaultDevice()->DestroyTexture(m_hTexture);
-
-  if (m_pContext != nullptr)
-  {
-    ezRmlUi::GetSingleton()->DeleteContext(m_pContext);
-    m_pContext = nullptr;
-  }
-
-  m_DataBindings.Clear();
-}
-
-void ezRmlUiCanvas3DComponent::OnActivated()
-{
-  SUPER::OnActivated();
-
-  GetOrCreateRmlContext()->ShowDocument();
-
-  // Update once to ensure correct initial state
-  Update();
-}
-
-void ezRmlUiCanvas3DComponent::OnDeactivated()
-{
-  m_pContext->HideDocument();
-
-  SUPER::OnDeactivated();
+  s >> m_vSize;
+  s >> m_bClearStaleInput;
+  s >> m_bIsInteractive;
+  s >> m_fDpiScale;
 }
 
 void ezRmlUiCanvas3DComponent::Update()
@@ -109,35 +93,19 @@ void ezRmlUiCanvas3DComponent::Update()
     }
   }
 
-  const ezTime tDiff = ezClock::GetGlobalClock()->GetTimeDiff();
-  m_bNeedsUpdate |= m_pContext->GetNextUpdateDelay() < ezMath::Max(tDiff.GetSeconds(), 1.0 / 240.0);
-
   m_bNeedsUpdate |= UpdateTexture();
 
-  for (auto& pDataBinding : m_DataBindings)
-  {
-    if (pDataBinding != nullptr)
-    {
-      m_bNeedsUpdate |= pDataBinding->Update();
-    }
-  }
-
-  if (m_bNeedsUpdate || m_bOnDemandUpdate == false)
-  {
-    m_pContext->Update();
-  }
-
-  m_bNeedsUpdate = false;
+  SUPER::Update();
 }
 
-void ezRmlUiCanvas3DComponent::ReceiveInput(const ezVec2& vMousePosInsideCanvas, ezRmlUiInputSnapshot input)
+bool ezRmlUiCanvas3DComponent::ReceiveInput(const ezVec2& vMousePosInsideCanvas, ezRmlUiInputSnapshot input)
 {
-  if (m_pContext == nullptr || !IsInteractive())
-    return;
-
-  m_InputProvider.Update(input);
-  m_bNeedsUpdate |= m_pContext->UpdateInput(vMousePosInsideCanvas, m_InputProvider);
-  m_iInputAge = 0;
+  if (IsInteractive() && SUPER::ReceiveInput(vMousePosInsideCanvas, input))
+  {
+    m_iInputAge = 0;
+    return true;
+  }
+  return false;
 }
 
 bool ezRmlUiCanvas3DComponent::RaycastInput(const ezVec3& vRayOrigin, const ezVec3& vRayDir, ezRmlUiInputSnapshot input)
@@ -177,8 +145,8 @@ bool ezRmlUiCanvas3DComponent::RaycastInput(const ezVec3& vRayOrigin, const ezVe
   }
 
   ezVec2 vCursorPos;
-  vCursorPos.x = static_cast<float>(m_vTextureSize.x) * vTexCoords.x;
-  vCursorPos.y = static_cast<float>(m_vTextureSize.y) * vTexCoords.y;
+  vCursorPos.x = static_cast<float>(m_vSize.x) * vTexCoords.x;
+  vCursorPos.y = static_cast<float>(m_vSize.y) * vTexCoords.y;
 
   ReceiveInput(vCursorPos, input);
 
@@ -242,34 +210,16 @@ bool ezRmlUiCanvas3DComponent::RaycastMeshTexCoords(const ezCpuMeshResource* pMe
   return false;
 }
 
-void ezRmlUiCanvas3DComponent::SetRmlResource(const ezRmlUiResourceHandle& hResource)
-{
-  if (m_hResource != hResource)
-  {
-    m_hResource = hResource;
-
-    if (m_pContext != nullptr)
-    {
-      if (m_pContext->LoadDocumentFromResource(m_hResource).Succeeded() && IsActive())
-      {
-        m_pContext->ShowDocument();
-      }
-
-      UpdateCachedValues();
-    }
-  }
-}
-
 void ezRmlUiCanvas3DComponent::SetTextureSize(const ezVec2U32& vSize)
 {
-  if (m_vTextureSize != vSize)
+  if (m_vSize != vSize)
   {
-    m_vTextureSize.x = ezMath::Min(vSize.x, 4096u);
-    m_vTextureSize.y = ezMath::Min(vSize.y, 4096u);
+    m_vSize.x = ezMath::Min(vSize.x, 4096u);
+    m_vSize.y = ezMath::Min(vSize.y, 4096u);
 
     if (m_pContext != nullptr)
     {
-      m_pContext->SetSize(m_vTextureSize);
+      m_pContext->SetSize(m_vSize);
     }
   }
 }
@@ -290,21 +240,6 @@ void ezRmlUiCanvas3DComponent::SetDpiScale(float fDpiScale)
   }
 }
 
-void ezRmlUiCanvas3DComponent::SetAutobindBlackboards(bool bAutobind)
-{
-  if (m_bAutobindBlackboards != bAutobind)
-  {
-    m_bAutobindBlackboards = bAutobind;
-
-    UpdateAutobinding();
-  }
-}
-
-void ezRmlUiCanvas3DComponent::SetOnDemandUpdate(bool bOnDemandUpdate)
-{
-  m_bOnDemandUpdate = bOnDemandUpdate;
-}
-
 void ezRmlUiCanvas3DComponent::SetClearStaleInput(bool bClearStaleInput)
 {
   m_bClearStaleInput = bClearStaleInput;
@@ -313,119 +248,6 @@ void ezRmlUiCanvas3DComponent::SetClearStaleInput(bool bClearStaleInput)
 void ezRmlUiCanvas3DComponent::SetInteractive(bool bIsInteractive)
 {
   m_bIsInteractive = bIsInteractive;
-}
-
-ezUInt32 ezRmlUiCanvas3DComponent::AddDataBinding(ezUniquePtr<ezRmlUiDataBinding>&& pDataBinding)
-{
-  // Document needs to be loaded again since data bindings have to be set before document load
-  if (m_pContext != nullptr)
-  {
-    if (pDataBinding->Initialize(*m_pContext).Succeeded())
-    {
-      if (m_pContext->LoadDocumentFromResource(m_hResource).Succeeded() && IsActive())
-      {
-        m_pContext->ShowDocument();
-      }
-    }
-  }
-
-  for (ezUInt32 i = 0; i < m_DataBindings.GetCount(); ++i)
-  {
-    if (pDataBinding == nullptr)
-    {
-      m_DataBindings[i] = std::move(pDataBinding);
-      return i;
-    }
-  }
-
-  ezUInt32 uiDataBindingIndex = m_DataBindings.GetCount();
-  m_DataBindings.PushBack(std::move(pDataBinding));
-  return uiDataBindingIndex;
-}
-
-void ezRmlUiCanvas3DComponent::RemoveDataBinding(ezUInt32 uiDataBindingIndex)
-{
-  auto& pDataBinding = m_DataBindings[uiDataBindingIndex];
-
-  if (m_pContext != nullptr)
-  {
-    pDataBinding->Deinitialize(*m_pContext);
-  }
-
-  m_DataBindings[uiDataBindingIndex] = nullptr;
-}
-
-ezUInt32 ezRmlUiCanvas3DComponent::AddBlackboardBinding(const ezSharedPtr<ezBlackboard>& pBlackboard)
-{
-  auto pDataBinding = EZ_DEFAULT_NEW(ezRmlUiInternal::BlackboardDataBinding, pBlackboard);
-  return AddDataBinding(pDataBinding);
-}
-
-void ezRmlUiCanvas3DComponent::RemoveBlackboardBinding(ezUInt32 uiDataBindingIndex)
-{
-  RemoveDataBinding(uiDataBindingIndex);
-}
-
-ezRmlUiContext* ezRmlUiCanvas3DComponent::GetOrCreateRmlContext()
-{
-  if (m_pContext != nullptr)
-  {
-    return m_pContext;
-  }
-
-  ezStringBuilder sName = "RmlUi_";
-  if (m_hResource.IsValid())
-  {
-    ezStringView sResourceID = m_hResource.GetResourceIdOrDescription();
-    sName.Append(sResourceID.GetFileName());
-  }
-  sName.AppendFormat("_{}", ezArgP(this));
-
-  m_pContext = ezRmlUi::GetSingleton()->CreateContext(sName, m_vTextureSize);
-
-  for (auto& pDataBinding : m_DataBindings)
-  {
-    pDataBinding->Initialize(*m_pContext).IgnoreResult();
-  }
-
-  m_pContext->LoadDocumentFromResource(m_hResource).IgnoreResult();
-
-  UpdateCachedValues();
-
-  return m_pContext;
-}
-
-void ezRmlUiCanvas3DComponent::SerializeComponent(ezWorldWriter& inout_stream) const
-{
-  SUPER::SerializeComponent(inout_stream);
-
-  ezStreamWriter& s = inout_stream.GetStream();
-
-  s << m_hResource;
-  s << m_vTextureSize;
-  s << m_bAutobindBlackboards;
-  s << m_bOnDemandUpdate;
-  s << m_bClearStaleInput;
-  s << m_bIsInteractive;
-  s << m_fDpiScale;
-}
-
-void ezRmlUiCanvas3DComponent::DeserializeComponent(ezWorldReader& inout_stream)
-{
-  SUPER::DeserializeComponent(inout_stream);
-  const ezUInt32 uiVersion = inout_stream.GetComponentTypeVersion(GetStaticRTTI());
-  ezStreamReader& s = inout_stream.GetStream();
-
-  s >> m_hResource;
-  s >> m_vTextureSize;
-  s >> m_bAutobindBlackboards;
-  s >> m_bOnDemandUpdate;
-  if (uiVersion >= 2)
-    s >> m_bClearStaleInput;
-  if (uiVersion >= 3)
-    s >> m_bIsInteractive;
-  if (uiVersion >= 4)
-    s >> m_fDpiScale;
 }
 
 ezResult ezRmlUiCanvas3DComponent::GetLocalBounds(ezBoundingBoxSphere& ref_bounds, bool& ref_bAlwaysVisible, ezMsgUpdateLocalBounds& ref_msg)
@@ -505,25 +327,14 @@ void ezRmlUiCanvas3DComponent::SetMesh(const ezMeshResourceHandle& hMesh)
   }
 }
 
-void ezRmlUiCanvas3DComponent::OnMsgReload(ezMsgRmlUiReload& msg)
-{
-  if (m_pContext != nullptr)
-  {
-    m_pContext->ReloadDocumentFromResource(m_hResource).IgnoreResult();
-    m_pContext->ShowDocument();
-
-    UpdateCachedValues();
-  }
-}
-
 bool ezRmlUiCanvas3DComponent::UpdateTexture()
 {
-  if (m_vTextureSize.x == 0 || m_vTextureSize.y == 0)
+  if (m_vSize.x == 0 || m_vSize.y == 0)
     return false;
 
   ezGALDevice* pDevice = ezGALDevice::GetDefaultDevice();
   const ezGALTexture* pTexture = pDevice->GetTexture(m_hTexture);
-  if (pTexture == nullptr || pTexture->GetDescription().m_uiWidth != m_vTextureSize.x || pTexture->GetDescription().m_uiHeight != m_vTextureSize.y)
+  if (pTexture == nullptr || pTexture->GetDescription().m_uiWidth != m_vSize.x || pTexture->GetDescription().m_uiHeight != m_vSize.y)
   {
     if (pTexture != nullptr)
     {
@@ -531,19 +342,19 @@ bool ezRmlUiCanvas3DComponent::UpdateTexture()
     }
 
     ezGALTextureCreationDescription desc;
-    desc.m_uiWidth = m_vTextureSize.x;
-    desc.m_uiHeight = m_vTextureSize.y;
+    desc.m_uiWidth = m_vSize.x;
+    desc.m_uiHeight = m_vSize.y;
     desc.m_Format = ezGALResourceFormat::RGBAUByteNormalized;
     desc.m_ResourceAccess.m_bImmutable = false;
-    if (ezMath::IsPowerOf2(m_vTextureSize.x) && ezMath::IsPowerOf2(m_vTextureSize.y))
+    if (ezMath::IsPowerOf2(m_vSize.x) && ezMath::IsPowerOf2(m_vSize.y))
     {
-      desc.m_uiMipLevelCount = ezMath::Max(ezMath::Log2i(m_vTextureSize.x), ezMath::Log2i(m_vTextureSize.y)) - 2;
+      desc.m_uiMipLevelCount = ezMath::Max(ezMath::Log2i(m_vSize.x), ezMath::Log2i(m_vSize.y)) - 2;
       desc.m_bAllowDynamicMipGeneration = true;
     }
 
     m_hTexture = pDevice->CreateTexture(desc);
 
-    m_pContext->SetSize(m_vTextureSize);
+    m_pContext->SetSize(m_vSize);
     m_pContext->SetDpiScale(m_fDpiScale);
 
     return true;
@@ -551,56 +362,5 @@ bool ezRmlUiCanvas3DComponent::UpdateTexture()
 
   return false;
 }
-
-void ezRmlUiCanvas3DComponent::UpdateCachedValues()
-{
-  m_ResourceEventUnsubscriber.Unsubscribe();
-
-  if (m_hResource.IsValid())
-  {
-    ezResourceLock pResource(m_hResource, ezResourceAcquireMode::PointerOnly);
-
-    pResource->m_ResourceEvents.AddEventHandler(
-      [hComponent = GetHandle(), pWorld = GetWorld()](const ezResourceEvent& e)
-      {
-        if (e.m_Type == ezResourceEvent::Type::ResourceContentUnloading)
-        {
-          pWorld->PostMessage(hComponent, ezMsgRmlUiReload(), ezTime::MakeZero());
-        }
-      },
-      m_ResourceEventUnsubscriber);
-  }
-}
-
-void ezRmlUiCanvas3DComponent::UpdateAutobinding()
-{
-  for (ezUInt32 uiIndex : m_AutoBindings)
-  {
-    RemoveDataBinding(uiIndex);
-  }
-
-  m_AutoBindings.Clear();
-
-  if (m_bAutobindBlackboards)
-  {
-    ezHybridArray<ezBlackboardComponent*, 4> blackboardComponents;
-
-    ezGameObject* pObject = GetOwner();
-    while (pObject != nullptr)
-    {
-      pObject->TryGetComponentsOfBaseType(blackboardComponents);
-
-      for (auto pBlackboardComponent : blackboardComponents)
-      {
-        pBlackboardComponent->EnsureInitialized();
-
-        m_AutoBindings.PushBack(AddBlackboardBinding(pBlackboardComponent->GetBoard()));
-      }
-
-      pObject = pObject->GetParent();
-    }
-  }
-}
-
 
 EZ_STATICLINK_FILE(RmlUiPlugin, RmlUiPlugin_Components_Implementation_RmlUiCanvas3DComponent);
